(PC-43137) refactor(app): replace GenericInfoPageIllustration by RemoteIllustration
